refactor: split reward routes into overlayAdminRewardMutations to reduce file complexity
